tier3/machine_learning_workers/requirements.txt
setuptools>=0
pytz>=0
zope.interface>=0
datetime>=4.2
enum34>=1.1.6
futures>=0
confluent-kafka>=0.11.5
pyscreeze>=0.1.18
pymsgbox>=1.0.6
PyTweening>=1.0.1
Pillow>=5.2.0
Xlib>=0.21
python3-xlib>=0
six>=1.10.0
pyautogui>=0.9.38
msgpack>=0
kafka-python>=0
redis>=0